improve testability / move init code out of context
