+++ databricks-kube/Cargo.toml
@@ -31,3 +31,6 @@ schemars = { version = "0.8.11", features = ["derive"] }
 tokio = { version = "1.19.2", features = ["macros", "rt-multi-thread"] }
 tokio-graceful-shutdown = "0.11.1"
 tokio-stream = "0.1.11"
+
+[dev-dependencies]
+tower-test = "0.4.0"
